Rename service to handler
Idea is that the handler only has to deal with the request parsing stuff
